wChat_AgentServer/pyproject.toml
[project]
name = "wchat-agent-server"
version = "0.1.0"
description = "wChat IM smart reply agent service"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.115,<0.117",
    "uvicorn[standard]>=0.32,<0.35",
    "pydantic>=2.9,<3",
    "pydantic-settings>=2.6,<3",
    "langgraph>=0.2.60,<0.3",
    "langchain-core>=0.3,<0.4",
    "openai>=1.54,<2",
    "pyyaml>=6.0,<7",
    "httpx>=0.27,<0.29",
    "sse-starlette>=2.1,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
    "pytest-asyncio>=0.24,<0.26",
    "respx>=0.21,<0.23",
    "ruff>=0.8,<0.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 110
target-version = "py312"
